# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.AuthSuccessResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
        type: string
      user:
        $ref: '#/definitions/handlers.AuthUserResponse'
    type: object
  handlers.AuthUserResponse:
    properties:
      avatar_url:
        example: https://cdn.example.com/avatar.png
        type: string
      display_name:
        example: Fafa
        type: string
      email:
        example: fafa@example.com
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      onboarding_completed_at:
        example: "2026-06-02T09:30:00Z"
        format: date-time
        type: string
      role:
        example: user
        type: string
      username:
        example: fafa
        type: string
      uuid:
        example: 018f6f6d-b0de-7b8f-bf91-43bc0b8f4c8a
        format: uuid
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      details:
        example: mail provider unavailable
        type: string
      error:
        example: 用户名和密码不匹配
        type: string
    type: object
...
  handlers.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  handlers.MessageResponse:
    properties:
      message:
        example: Verification code sent
        type: string
    type: object
...
  handlers.RegisterInput:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      password_confirm:
        type: string
      username:
        type: string
      verification_code:
        type: string
    required:
    - email
    - password
    - password_confirm
    - username
    - verification_code
    type: object
...
  handlers.SendVerificationInput:
    properties:
      email:
        type: string
      turnstile_token:
        type: string
    required:
    - email
    type: object
...
  handlers.VerifyEmailInput:
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
...
info:
  contact: {}
  description: Atoman 后端 API 文档。
  title: Atoman API
  version: "1.0"
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: 使用用户名或邮箱登录并返回登录态。
      parameters:
      - description: 登录请求
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 用户登录
      tags:
      - auth
...
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: 验证邮箱验证码后创建账号并返回登录态。
      parameters:
      - description: 注册请求
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.AuthSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 注册新用户
      tags:
      - auth
  /api/v1/auth/send-verification:
    post:
      consumes:
      - application/json
      description: 向指定邮箱发送 6 位验证码。
      parameters:
      - description: 验证码请求
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.SendVerificationInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 发送邮箱验证码
      tags:
      - auth
  /api/v1/auth/session:
    get:
      description: 读取当前登录用户信息。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: 获取当前会话
      tags:
      - auth
  /api/v1/auth/verify-email:
    post:
      consumes:
      - application/json
      description: 校验邮箱与验证码是否匹配。
      parameters:
      - description: 邮箱验证请求
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.VerifyEmailInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 校验邮箱验证码
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 使用 Bearer Token，例如：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    description: 使用登录后写入的 atoman_token Cookie
    in: cookie
    name: atoman_token
    type: apiKey
swagger: "2.0"